+++ lib/mini_suffix.rb
@@ -5,9 +5,10 @@ module MiniSuffix
   extend FFI::Library
 
   ffi_lib [
-    File.join(File.dirname(File.expand_path('..', __FILE__)), 'vendor/libpsl.so'),
+    File.join(File.dirname(File.expand_path('..', __FILE__)), 'vendor/libpsl.x86_64.so'),  
     File.join(File.dirname(File.expand_path('..', __FILE__)), 'vendor/libpsl.ppc64le.so'),
-    File.join(File.dirname(File.expand_path('..', __FILE__)), 'vendor/libpsl.dylib'),
+    File.join(File.dirname(File.expand_path('..', __FILE__)), 'vendor/libpsl.darwin-i386.dylib'),
+    File.join(File.dirname(File.expand_path('..', __FILE__)), 'vendor/libpsl.darwin-arm64.dylib'),
   ]
 
   attach_function :psl_builtin, [], :pointer