+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# Unreleased
+# 0.3.2
+
+* Bump libpsl to v0.21.1, and bump binaries
+* Add binary for Apple silicon devices
+
+  *davidtaylorhq*
+
+# 0.3.1
 
 * Add libpsl shared libary for Linux ppc64le